requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0.1
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
jwt==1.4.0
mysqlclient==2.2.7
pillow==12.1.0
pycparser==2.23
PyJWT==2.10.1
qrcode==8.2
sqlparse==0.5.5
